backend/requirements.txt
# Updated for Python 3.13 compatibility (uses pre-built wheels)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.23
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-multipart>=0.0.6
python-dateutil>=2.8.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Optional: Uncomment for PostgreSQL support (requires pg_config in PATH)
# psycopg2-binary==2.9.9
